About NycBikeSafeRide
The goal of our project is to make bike riders more aware of crashes within the various community districts where the citibikes are located. This would help cyclist avoid accidents based on bike lanes and the routes they choose to cycle throughout the city. Our website provides riders with useful information such as the address and location of all Citibikes station.
Additionally, the rider is able to receive a pre ride estimation of the mileage and duration of their chosen route. Furthermore, our website provides its users with the crash charts which they can use as a tool to avoid certain crash areas. The user can easily take a look at the charts and get a better idea of how safe or unsafe a community is and the amount of bike crashes that have occurred in that area.
Technology Used - Php, PostgrqSql, Javascript, Bootstrap